+++ relo/src/main/java/com/my/relo/service/StyleService.java
@@ -1,7 +1,9 @@
 package com.my.relo.service;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -9,11 +11,16 @@ import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.my.relo.dto.LikesDTO;
+import com.my.relo.dto.MemberDTO;
 import com.my.relo.dto.ReplyDTO;
 import com.my.relo.dto.StyleDTO;
 import com.my.relo.dto.StyleTagDTO;
+import com.my.relo.entity.Likes;
+import com.my.relo.entity.Member;
+import com.my.relo.entity.Reply;
 import com.my.relo.entity.Style;
 import com.my.relo.entity.StyleTag;
 import com.my.relo.entity.StyleTagEmbedded;
@@ -36,18 +43,20 @@ public class StyleService {
 	private LikesRepository lr;
 	
 	@Autowired
-	private ReplyService rs;
-	
-	@Autowired
 	private LikesService ls;
 	
 	@Autowired
 	ModelMapper modelMapper;
 	
-	//게시판 추가 
+	/**
+	 * 글 작성 
+	 * @param styleDTO
+	 * @throws AddException
+	 */
 	public void write(StyleDTO styleDTO) throws AddException{
 		List<StyleTagDTO> tagList = styleDTO.getTagList();
-		Style style = styleDTO.toEntity();
+		MemberDTO m = styleDTO.getMember();
+		Style style = styleDTO.toEntity(m);
 		sr.save(style);
 		Optional<Style> optS1 = sr.findById(style.getStyleNum());
 		Style s = optS1.get();
@@ -59,14 +68,26 @@ public class StyleService {
 			str.save(tag);
 		}
 	}
-	
-	//게시판 삭제
-	public void deleteByStyleNum(Long styleNum) throws RemoveException{
+	/**
+	 * 게시판 삭제 후 나머지 댓글, 좋아요, 해시태그 삭제 
+	 * @param styleNum
+	 * @throws RemoveException
+	 * @throws FindException 
+	 */
+	public void deleteByStyleNum(Long styleNum) throws RemoveException, FindException{
+		
+		Optional<Style> optA= sr.findById(styleNum);
+		if(!optA.isPresent()) {
+			throw new FindException("존재하지 않는 게시물입니다.");
+		}
 		lr.deleteByStyleNumList(styleNum);
 		sr.deleteById(styleNum);
 	}
-	
-	//게시판 최신순 리스트 출력
+	/**
+	 * 최신순 리스트 출력 
+	 * @return
+	 * @throws FindException
+	 */
 	public List<StyleDTO> listByStyleNum() throws FindException{
 		List<Style> styleList = sr.findAll(Sort.by(Sort.Direction.DESC,"styleNum"));
 		List<StyleDTO> styleDTOList = 
@@ -79,8 +100,12 @@ public class StyleService {
 		}
 		return list;
 	}
-	
-	//게시판 해시태그별 리스트 출력
+	/**
+	 * 해시태그 별 리스트 출력 
+	 * @param hashName
+	 * @return
+	 * @throws FindException
+	 */
 	public List<StyleDTO> listByHashName(String hashName) throws FindException{
 		List<Style> styleList = sr.listByHashName(hashName);
 		List<StyleDTO> styleDTOList = 
@@ -94,7 +119,11 @@ public class StyleService {
 		return list;
 	}
 	
-	//게시판 좋아요순 리스트 출력
+	/**
+	 * 좋아요 리스트 출력 
+	 * @return List<StyleDTO>
+	 * @throws FindException
+	 */
 	public List<StyleDTO> listByLikes() throws FindException{
 		List<Style> styleList = sr.listByLikes();
 		List<StyleDTO> styleDTOList = 
@@ -108,7 +137,11 @@ public class StyleService {
 		return list;
 	}
 	
-	//게시판 조회수순 리스트 출력
+	/**
+	 * 조회수순 리스트 출력 
+	 * @return List<StyleDTO>
+	 * @throws FindException
+	 */
 	public List<StyleDTO> listBystyleCnt() throws FindException{
 		List<Style> styleList = sr.findAll(Sort.by(Sort.Direction.DESC,"styleCnt","styleNum"));
 		List<StyleDTO> styleDTOList = 
@@ -122,7 +155,12 @@ public class StyleService {
 		return list;
 	}
 	
-	//내가 쓴 게시판 리스트 출력
+	/**
+	 * 내가 쓴 리스트 출력 
+	 * @param mNum 
+	 * @return
+	 * @throws FindException
+	 */
 	public List<StyleDTO> listByMNum(Long mNum) throws FindException{
 		List<Style> styleList = sr.findBymNum(mNum);
 		List<StyleDTO> styleDTOList =
@@ -136,24 +174,72 @@ public class StyleService {
 		return list;
 	}
 	
-	//게시판 상세보기 출력
+	/**
+	 * 게시판 상세보기 출력 
+	 * @param styleNum
+	 * @return
+	 * @throws FindException
+	 */
+	@Transactional
 	public StyleDTO styleDetail(Long styleNum) throws FindException{
-		Optional<Style> optS = sr.findById(styleNum);
-		List <LikesDTO> likes = ls.listByStyleNum(styleNum);
-		List <ReplyDTO> repList = rs.listByStyleNum(styleNum);
-		Style style = optS.get();
-		StyleDTO styleDTO = 
-				modelMapper.map(style, StyleDTO.class);
-		styleDTO.setLikesList(likes);
-		styleDTO.setReplyList(repList);
 		
-		return styleDTO;
+		Optional<Style> optS = sr.findById(styleNum);
+		Style style = optS.get();
+		StyleDTO sDTO = new StyleDTO();
+		MemberDTO mDTO1 = new MemberDTO();
+		mDTO1.setMNum(style.getMember().getMNum());
+		mDTO1.setId(style.getMember().getId());
+		sDTO.setMember(mDTO1);
+		sDTO.setStyleNum(style.getStyleNum());
+		sDTO.setStyleDate(style.getStyleDate());
+		
+		List<Reply> list= style.getRepList();
+		List<ReplyDTO> listDTO = new ArrayList<>();
+		for(Reply r: list) {
+			ReplyDTO rDTO = new ReplyDTO();
+			
+			Member m = r.getMember();
+			MemberDTO mDTO = new MemberDTO();
+			mDTO.setId(m.getId());
+			
+			rDTO.setRepNum(r.getRepNum());
+			rDTO.setRepContent(r.getRepContent());
+			rDTO.setRepDate(r.getRepDate());
+			rDTO.setMember(mDTO);
+			
+			listDTO.add(rDTO);
+		}
+		
+		sDTO.setReplyList(listDTO);
+		sDTO.setStyleCnt(style.getStyleCnt());
+
+		List<Likes> likeEntityList = style.getLikeList();
+		List<LikesDTO> likesDTOList = 
+				likeEntityList.stream().map(likes -> modelMapper.map(likes, LikesDTO.class)).collect(Collectors.toList());
+		
+		sDTO.setLikesList(likesDTOList);
+		
+		List<StyleTag> tagEntityList = style.getTagList();
+		List<StyleTagDTO> tagDTOList = 
+				tagEntityList.stream().map(tags -> modelMapper.map(tags, StyleTagDTO.class)).collect((Collectors.toList()));
+		
+		sDTO.setTagList(tagDTOList);
+
+		return sDTO;
 	}
 	
-	//게시판 수정 
-	public void edit(List<StyleTagDTO> list) throws AddException{
-		StyleTagDTO st= list.get(0);
-		Long styleNum = st.getSte().getStyleNum();
+	/**
+	 * 게시판 수정
+	 * @param StyleDTO
+	 * @throws AddException
+	 * @throws FindException 
+	 */
+	public void edit(StyleDTO styleDTO) throws AddException, FindException{
+		List <StyleTagDTO> list = styleDTO.getTagList();
+		if(list == null) {
+			throw new FindException("존재하지 않는 게시물입니다.");
+		}
+		Long styleNum = styleDTO.getStyleNum();
 		str.deleteByStyleNum(styleNum);
 		Optional<Style> optS= sr.findById(styleNum);
 		Style s = optS.get();
@@ -166,7 +252,11 @@ public class StyleService {
 		}
 	}
 	
-	//조회수 증가
+	/**
+	 * 상세조회 클릭시 조회수 증가 
+	 * @param styleNum
+	 * @throws AddException
+	 */
 	public void plusCnt(Long styleNum) throws AddException{
 		sr.updateCnt(styleNum);
 	}
